third_party/blink/web_tests/virtual/threaded/compositing/element-animate-positive-delay.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div id="target">x</div>
<script>
promise_test(function() {
  var animation = target.animate({
    transform: ['rotate(0deg)', 'rotate(180deg)'],
  }, {
    duration: 100000,
    delay: 1,
  });
  return animation.ready.then(() => {
    assert_true(internals.isCompositedAnimation(animation));
  });
}, 'element.animate() with a positive delay should run compositor animations');
</script>
